+++ src/build.rs
@@ -530,9 +530,9 @@ where
     ///
     /// let m = mock([
     ///     // the first call MUST be method(1) and it will return "a"
-    ///     Trait__method.next_call(matching!(1)).returns_static("a").once().in_order(),
+    ///     TraitMock::method.next_call(matching!(1)).returns_static("a").once().in_order(),
     ///     // the second call MUST be method(2) and it will return "b"
-    ///     Trait__method.next_call(matching!(2)).returns_static("b").once().in_order(),
+    ///     TraitMock::method.next_call(matching!(2)).returns_static("b").once().in_order(),
     ///     // there may be no more calls to this mock, as it has no stubs in it
     /// ]);
     ///